backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
gunicorn
django-cors-headers
django-filter
whitenoise